backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai>=1.12.0
python-dotenv==1.0.0
feedparser>=6.0.11
geopy==2.4.1
# Optional: newspaper3k (may have compatibility issues with Python 3.13)
# If installation fails, the app will work without it using RSS feed data only
# newspaper3k==0.2.8
# nltk==3.8.1
